refactor: Extract util method from view
- Extracted method from the view, since we can use the same method in the UIView player.
